+++ src/app/dashboard/(active)/breadcrumb.tsx
@@ -1,6 +1,7 @@
 "use client"
 import { usePathname } from "next/navigation"
 import { Fragment, useMemo } from "react"
+import { NAV_MAP } from "@/components/dashboard-sidebar/data"
 import {
   BreadcrumbItem,
   BreadcrumbLink,
@@ -9,11 +10,15 @@ import {
   Breadcrumb as BreadcrumbRoot,
   BreadcrumbSeparator,
 } from "@/components/ui/breadcrumb"
-import { getBreadcrumbs } from "@/components/dashboard-sidebar/main-nav"
+
+export type BreadcrumbItem = {
+  title: string
+  url?: string
+}
 
 export function Breadcrumb() {
   const pathname = usePathname()
-  const items = useMemo(() => getBreadcrumbs(pathname), [pathname])
+  const items = useMemo(() => getBreadcrumbs(NAV_MAP, pathname), [pathname])
 
   return (
     <BreadcrumbRoot>
@@ -35,4 +40,31 @@ export function Breadcrumb() {
       </BreadcrumbList>
     </BreadcrumbRoot>
   )
+}
+
+function getBreadcrumbs(navMap: Map<string, string>, pathname: string): BreadcrumbItem[] {
+  const segments = pathname.split("/").filter(Boolean)
+  const breadcrumbs: BreadcrumbItem[] = []
+
+  let currentPath = ""
+
+  for (const segment of segments) {
+    currentPath += `/${segment}`
+    let title = navMap.get(currentPath)
+    if (!title) {
+      if (isUUIDorId(segment)) {
+        title = "Details"
+      } else {
+        title = segment.charAt(0).toUpperCase() + segment.slice(1)
+      }
+    }
+
+    breadcrumbs.push({ title, url: currentPath })
+  }
+
+  return breadcrumbs
+}
+
+function isUUIDorId(segment: string) {
+  return !isNaN(Number(segment)) || segment.length > 20 // Regex custom a seconda dei tuoi ID
 }
